backend/requirements.txt
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
requests
elasticsearch

